+++ tests/test_cli_core.py
@@ -1,3 +1,7 @@
+from typing import Optional
+
+import pytest
+
 from _incydr_cli.main import incydr
 
 
@@ -9,3 +13,27 @@ def test_cli_auth_missing_error_prints_missing_vars(runner, monkeypatch):
     assert "INCYDR_API_CLIENT_SECRET" in result.output
     assert "INCYDR_URL" not in result.output
     assert "INCYDR_API_CLIENT_ID" not in result.output
+
+
+@pytest.mark.disable_autouse
+def test_cli_user_agent(runner, httpserver_auth):
+    def starts_with_matcher(
+        header_name: str, actual: Optional[str], expected: str
+    ) -> bool:
+        if actual is None:
+            return False
+
+        return actual.startswith(expected)
+
+    httpserver_auth.expect_ordered_request(
+        "/v1/users",
+        method="GET",
+        headers={"User-Agent": "incydrCLI"},
+        header_value_matcher=starts_with_matcher,
+    ).respond_with_json({"users": [], "totalCount": 0})
+    result = runner.invoke(
+        incydr, ["users", "list", "--log-stderr", "--log-level", "DEBUG"]
+    )
+    httpserver_auth.check()
+    assert result.exit_code == 0
+    assert "No results found" in result.output